+++ .settings/org.eclipse.wst.common.component
@@ -1,32 +1,39 @@
 <?xml version="1.0" encoding="UTF-8"?><project-modules id="moduleCoreId" project-version="1.5.0">
-            	
+                	
+    
     
     
     <wb-module deploy-name="ASM-JAVA4-0.0.1-SNAPSHOT">
-                        		
+                                		
+        
         
         
         <wb-resource deploy-path="/" source-path="/target/m2e-wtp/web-resources"/>
-                        		
+                                		
+        
         
         
         <wb-resource deploy-path="/" source-path="/src/main/webapp" tag="defaultRootSource"/>
-                        		
+                                		
+        
         
         
         <wb-resource deploy-path="/WEB-INF/classes" source-path="/src/main/java"/>
-        <wb-resource deploy-path="/WEB-INF/classes" source-path="/src/test/java"/>
-                        		
+                                		
+        
         
         
         <property name="java-output-path" value="/ASM-JAVA4/build/classes"/>
-                
+                        
+        
         <property name="context-root" value="ASM-JAVA4"/>
-                        	
+                                	
+    
     
     
     </wb-module>
-            
+                
+
 
 
 </project-modules>
