+++ Resources/Data_Khanh/datafile.properties
@@ -1,3 +1,1 @@
-url=https://phptravels.net/login
-username=user@phptravels.com
-passsword=demouser
+url=http://localhost:8081/ASM-JAVA4/Homepage
