+++ Resources/Data_Khanh/locator.properties
@@ -1,4 +1,3 @@
-Username_field = name:email
-Password_field = name:password
-Login_button = classname:btn-lg
-online_user = classname:RTL
+
+next_button = classname:slick-next
+prev_button = classname:slick-prev
