+++ src/main/webapp/user/layout_home/slider.jsp
@@ -70,7 +70,6 @@
 					</div>
 				</div>
 			</div>
-			
 		</div>
 	</div>
 </div>
