+++ test-output/Default suite/Default test.html
@@ -57,9 +57,9 @@ function toggleAllBoxes() {
 <tr>
 <td>Tests passed/Failed/Skipped:</td><td>3/0/0</td>
 </tr><tr>
-<td>Started on:</td><td>Tue Nov 22 11:47:04 ICT 2022</td>
+<td>Started on:</td><td>Wed Nov 23 00:01:41 ICT 2022</td>
 </tr>
-<tr><td>Total time:</td><td>30 seconds (30188 ms)</td>
+<tr><td>Total time:</td><td>13 seconds (13428 ms)</td>
 </tr><tr>
 <td>Included groups:</td><td></td>
 </tr><tr>
@@ -75,20 +75,20 @@ function toggleAllBoxes() {
 <td><b>Instance</b></td>
 </tr>
 <tr>
-<td title='test.Han.TestHan.DoLogin()'><b>DoLogin</b><br>Test class: test.Han.TestHan<br>Test method: Perform Login</td>
-<td></td>
-<td>16</td>
-<td>test.Han.TestHan@1bd4fdd</td></tr>
-<tr>
-<td title='test.Han.TestHan.FillLoginDetails()'><b>FillLoginDetails</b><br>Test class: test.Han.TestHan<br>Test method: Fill the Login Details</td>
+<td title='test.Khanh.TestSlider.Button_Detail()'><b>Button_Detail</b><br>Test class: test.Khanh.TestSlider<br>Test method: Clik the button more detail on slider</td>
 <td></td>
 <td>1</td>
-<td>test.Han.TestHan@1bd4fdd</td></tr>
+<td>test.Khanh.TestSlider@4f83df68</td></tr>
 <tr>
-<td title='test.Han.TestHan.lauchWebsite()'><b>lauchWebsite</b><br>Test class: test.Han.TestHan<br>Test method: Open TestNG Demo Website for Login Test</td>
+<td title='test.Khanh.TestSlider.Button_Slide()'><b>Button_Slide</b><br>Test class: test.Khanh.TestSlider<br>Test method: Clik the button on slider</td>
 <td></td>
-<td>7</td>
-<td>test.Han.TestHan@1bd4fdd</td></tr>
+<td>2</td>
+<td>test.Khanh.TestSlider@4f83df68</td></tr>
+<tr>
+<td title='test.Khanh.TestSlider.lauchWebsite()'><b>lauchWebsite</b><br>Test class: test.Khanh.TestSlider<br>Test method: Open TestNG Demo Website</td>
+<td></td>
+<td>5</td>
+<td>test.Khanh.TestSlider@4f83df68</td></tr>
 </table><p>
 </body>
 </html>
