+++ test-output/Default suite/Default test.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- Generated by org.testng.reporters.JUnitXMLReporter -->
-<testsuite ignored="0" hostname="HAN-NGUYEN" failures="0" tests="3" name="Default test" time="30.188" errors="0" timestamp="2022-11-22T11:47:35 ICT">
-  <testcase classname="test.Han.TestHan" name="lauchWebsite" time="7.808"/>
-  <testcase classname="test.Han.TestHan" name="FillLoginDetails" time="1.318"/>
-  <testcase classname="test.Han.TestHan" name="DoLogin" time="16.339"/>
+<testsuite ignored="0" hostname="KHANH-PC" failures="0" tests="3" name="Default test" time="13.428" errors="0" timestamp="2022-11-23T00:01:55 ICT">
+  <testcase classname="test.Khanh.TestSlider" name="lauchWebsite" time="5.558"/>
+  <testcase classname="test.Khanh.TestSlider" name="Button_Slide" time="2.304"/>
+  <testcase classname="test.Khanh.TestSlider" name="Button_Detail" time="1.364"/>
 </testsuite> <!-- Default test -->
