+++ test-output/emailable-report.html
@@ -9,12 +9,12 @@
 <table>
 <tr><th>Test</th><th># Passed</th><th># Skipped</th><th># Retried</th><th># Failed</th><th>Time (ms)</th><th>Included Groups</th><th>Excluded Groups</th></tr>
 <tr><th colspan="8">Default suite</th></tr>
-<tr><td><a href="#t0">Default test</a></td><td class="num">3</td><td class="num">0</td><td class="num">0</td><td class="num">0</td><td class="num">30,188</td><td></td><td></td></tr>
+<tr><td><a href="#t0">Default test</a></td><td class="num">3</td><td class="num">0</td><td class="num">0</td><td class="num">0</td><td class="num">13,428</td><td></td><td></td></tr>
 </table>
-<table id='summary'><thead><tr><th>Class</th><th>Method</th><th>Start</th><th>Time (ms)</th></tr></thead><tbody><tr><th colspan="4">Default suite</th></tr></tbody><tbody id="t0"><tr><th colspan="4">Default test &#8212; passed</th></tr><tr class="passedeven"><td rowspan="3">test.Han.TestHan</td><td><a href="#m0">DoLogin</a></td><td rowspan="1">1669092438505</td><td rowspan="1">16339</td></tr><tr class="passedeven"><td><a href="#m1">FillLoginDetails</a></td><td rowspan="1">1669092437185</td><td rowspan="1">1318</td></tr><tr class="passedeven"><td><a href="#m2">lauchWebsite</a></td><td rowspan="1">1669092429371</td><td rowspan="1">7808</td></tr></tbody>
+<table id='summary'><thead><tr><th>Class</th><th>Method</th><th>Start</th><th>Time (ms)</th></tr></thead><tbody><tr><th colspan="4">Default suite</th></tr></tbody><tbody id="t0"><tr><th colspan="4">Default test &#8212; passed</th></tr><tr class="passedeven"><td rowspan="3">test.Khanh.TestSlider</td><td><a href="#m0">Button_Detail</a></td><td rowspan="1">1669136512794</td><td rowspan="1">1364</td></tr><tr class="passedeven"><td><a href="#m1">Button_Slide</a></td><td rowspan="1">1669136510487</td><td rowspan="1">2304</td></tr><tr class="passedeven"><td><a href="#m2">lauchWebsite</a></td><td rowspan="1">1669136504918</td><td rowspan="1">5558</td></tr></tbody>
 </table>
-<h2>Default test</h2><h3 id="m0">test.Han.TestHan#DoLogin</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
-<h3 id="m1">test.Han.TestHan#FillLoginDetails</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
-<h3 id="m2">test.Han.TestHan#lauchWebsite</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
+<h2>Default test</h2><h3 id="m0">test.Khanh.TestSlider#Button_Detail</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
+<h3 id="m1">test.Khanh.TestSlider#Button_Slide</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
+<h3 id="m2">test.Khanh.TestSlider#lauchWebsite</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
 </body>
 </html>
