+++ test-output/junitreports/TEST-test.Khanh.TestMau.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Generated by org.testng.reporters.JUnitReportReporter -->
+<testsuite hostname="KHANH-PC" failures="0" tests="3" name="test.Khanh.TestMau" time="8.385" errors="0" timestamp="2022-11-22T23:56:15 ICT" skipped="0">
+  <testcase classname="test.Khanh.TestMau" name="lauchWebsite" time="4.832"/>
+  <system-out/>
+  <testcase classname="test.Khanh.TestMau" name="Button_Slide" time="2.381"/>
+  <system-out/>
+  <testcase classname="test.Khanh.TestMau" name="Button_Detail" time="1.172"/>
+  <system-out/>
+</testsuite> <!-- test.Khanh.TestMau -->
