+++ test-output/junitreports/TEST-test.Khanh.TestSlider.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Generated by org.testng.reporters.JUnitReportReporter -->
+<testsuite hostname="KHANH-PC" failures="0" tests="3" name="test.Khanh.TestSlider" time="9.226" errors="0" timestamp="2022-11-23T00:01:55 ICT" skipped="0">
+  <testcase classname="test.Khanh.TestSlider" name="lauchWebsite" time="5.558"/>
+  <system-out/>
+  <testcase classname="test.Khanh.TestSlider" name="Button_Slide" time="2.304"/>
+  <system-out/>
+  <testcase classname="test.Khanh.TestSlider" name="Button_Detail" time="1.364"/>
+  <system-out/>
+</testsuite> <!-- test.Khanh.TestSlider -->
