+++ test-output/old/Default suite/classes.html
@@ -4,22 +4,22 @@
 <th>Method name</th>
 <th>Groups</th>
 </tr><tr>
-<td>test.Han.TestHan</td>
+<td>test.Khanh.TestSlider</td>
 <td>&nbsp;</td><td>&nbsp;</td></tr>
 <tr>
 <td align='center' colspan='3'>@Test</td>
 </tr>
 <tr>
 <td>&nbsp;</td>
-<td>DoLogin</td>
+<td>Button_Detail</td>
+<td>&nbsp;</td></tr>
+<tr>
+<td>&nbsp;</td>
+<td>Button_Slide</td>
 <td>&nbsp;</td></tr>
 <tr>
 <td>&nbsp;</td>
 <td>lauchWebsite</td>
-<td>&nbsp;</td></tr>
-<tr>
-<td>&nbsp;</td>
-<td>FillLoginDetails</td>
 <td>&nbsp;</td></tr>
 <tr>
 <td align='center' colspan='3'>@BeforeClass</td>
