+++ test-output/old/Default suite/methods-alphabetical.html
@@ -1,14 +1,14 @@
 <h2>Methods run, sorted chronologically</h2><h3>&gt;&gt; means before, &lt;&lt; means after</h3><p/><br/><em>Default suite</em><p/><small><i>(Hover the method name to see the test class name)</i></small><p/>
 <table border="1">
 <tr><th>Time</th><th>Delta (ms)</th><th>Suite<br>configuration</th><th>Test<br>configuration</th><th>Class<br>configuration</th><th>Groups<br>configuration</th><th>Method<br>configuration</th><th>Test<br>method</th><th>Thread</th><th>Instances</th></tr>
-<tr bgcolor="c386c8">  <td>22/11/22 11:47:18</td>   <td>0</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="TestHan.DoLogin()[pri:3, instance:test.Han.TestHan@1bd4fdd]">DoLogin</td> 
-  <td>main@640070680</td>   <td></td> </tr>
-<tr bgcolor="c386c8">  <td>22/11/22 11:47:17</td>   <td>-1320</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="TestHan.FillLoginDetails()[pri:2, instance:test.Han.TestHan@1bd4fdd]">FillLoginDetails</td> 
-  <td>main@640070680</td>   <td></td> </tr>
-<tr bgcolor="c386c8">  <td>22/11/22 11:47:09</td>   <td>-9134</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="TestHan.lauchWebsite()[pri:1, instance:test.Han.TestHan@1bd4fdd]">lauchWebsite</td> 
-  <td>main@640070680</td>   <td></td> </tr>
-<tr bgcolor="c386c8">  <td>22/11/22 11:47:05</td>   <td>-13503</td> <td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;TestHan.suiteSetUp()[pri:0, instance:test.Han.TestHan@1bd4fdd]">&gt;&gt;suiteSetUp</td> 
-<td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td>  <td>main@640070680</td>   <td></td> </tr>
-<tr bgcolor="c386c8">  <td>22/11/22 11:47:34</td>   <td>16343</td> <td>&nbsp;</td><td>&nbsp;</td><td title="&lt;&lt;TestHan.suiteTearDown()[pri:0, instance:test.Han.TestHan@1bd4fdd]">&lt;&lt;suiteTearDown</td> 
-<td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td>  <td>main@640070680</td>   <td></td> </tr>
+<tr bgcolor="6decaa">  <td>22/11/23 00:01:52</td>   <td>0</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="TestSlider.Button_Detail()[pri:3, instance:test.Khanh.TestSlider@4f83df68]">Button_Detail</td> 
+  <td>main@1995265320</td>   <td></td> </tr>
+<tr bgcolor="6decaa">  <td>22/11/23 00:01:50</td>   <td>-2307</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="TestSlider.Button_Slide()[pri:2, instance:test.Khanh.TestSlider@4f83df68]">Button_Slide</td> 
+  <td>main@1995265320</td>   <td></td> </tr>
+<tr bgcolor="6decaa">  <td>22/11/23 00:01:44</td>   <td>-7876</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="TestSlider.lauchWebsite()[pri:1, instance:test.Khanh.TestSlider@4f83df68]">lauchWebsite</td> 
+  <td>main@1995265320</td>   <td></td> </tr>
+<tr bgcolor="6decaa">  <td>22/11/23 00:01:41</td>   <td>-10998</td> <td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;TestSlider.suiteSetUp()[pri:0, instance:test.Khanh.TestSlider@4f83df68]">&gt;&gt;suiteSetUp</td> 
+<td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td>  <td>main@1995265320</td>   <td></td> </tr>
+<tr bgcolor="6decaa">  <td>22/11/23 00:01:54</td>   <td>1376</td> <td>&nbsp;</td><td>&nbsp;</td><td title="&lt;&lt;TestSlider.suiteTearDown()[pri:0, instance:test.Khanh.TestSlider@4f83df68]">&lt;&lt;suiteTearDown</td> 
+<td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td>  <td>main@1995265320</td>   <td></td> </tr>
 </table>
