+++ test-output/testng-results.xml
@@ -2,32 +2,32 @@
 <testng-results ignored="0" total="3" passed="3" failed="0" skipped="0">
   <reporter-output>
   </reporter-output>
-  <suite started-at="2022-11-22T11:47:04 ICT" name="Default suite" finished-at="2022-11-22T11:47:35 ICT" duration-ms="30188">
+  <suite started-at="2022-11-23T00:01:41 ICT" name="Default suite" finished-at="2022-11-23T00:01:55 ICT" duration-ms="13428">
     <groups>
     </groups>
-    <test started-at="2022-11-22T11:47:04 ICT" name="Default test" finished-at="2022-11-22T11:47:35 ICT" duration-ms="30188">
-      <class name="test.Han.TestHan">
-        <test-method is-config="true" signature="suiteSetUp()[pri:0, instance:test.Han.TestHan@1bd4fdd]" started-at="2022-11-22T11:47:04 ICT" name="suiteSetUp" finished-at="2022-11-22T11:47:09 ICT" duration-ms="4362" status="PASS">
+    <test started-at="2022-11-23T00:01:41 ICT" name="Default test" finished-at="2022-11-23T00:01:55 ICT" duration-ms="13428">
+      <class name="test.Khanh.TestSlider">
+        <test-method is-config="true" signature="suiteSetUp()[pri:0, instance:test.Khanh.TestSlider@4f83df68]" started-at="2022-11-23T00:01:41 ICT" name="suiteSetUp" finished-at="2022-11-23T00:01:44 ICT" duration-ms="3100" status="PASS">
           <reporter-output>
           </reporter-output>
         </test-method> <!-- suiteSetUp -->
-        <test-method signature="lauchWebsite()[pri:1, instance:test.Han.TestHan@1bd4fdd]" started-at="2022-11-22T11:47:09 ICT" name="lauchWebsite" description="Open TestNG Demo Website for Login Test" finished-at="2022-11-22T11:47:17 ICT" duration-ms="7808" status="PASS">
+        <test-method signature="lauchWebsite()[pri:1, instance:test.Khanh.TestSlider@4f83df68]" started-at="2022-11-23T00:01:44 ICT" name="lauchWebsite" description="Open TestNG Demo Website" finished-at="2022-11-23T00:01:50 ICT" duration-ms="5558" status="PASS">
           <reporter-output>
           </reporter-output>
         </test-method> <!-- lauchWebsite -->
-        <test-method signature="FillLoginDetails()[pri:2, instance:test.Han.TestHan@1bd4fdd]" started-at="2022-11-22T11:47:17 ICT" name="FillLoginDetails" description="Fill the Login Details" finished-at="2022-11-22T11:47:18 ICT" duration-ms="1318" status="PASS">
+        <test-method signature="Button_Slide()[pri:2, instance:test.Khanh.TestSlider@4f83df68]" started-at="2022-11-23T00:01:50 ICT" name="Button_Slide" description="Clik the button on slider" finished-at="2022-11-23T00:01:52 ICT" duration-ms="2304" status="PASS">
           <reporter-output>
           </reporter-output>
-        </test-method> <!-- FillLoginDetails -->
-        <test-method signature="DoLogin()[pri:3, instance:test.Han.TestHan@1bd4fdd]" started-at="2022-11-22T11:47:18 ICT" name="DoLogin" description="Perform Login" finished-at="2022-11-22T11:47:34 ICT" duration-ms="16339" status="PASS">
+        </test-method> <!-- Button_Slide -->
+        <test-method signature="Button_Detail()[pri:3, instance:test.Khanh.TestSlider@4f83df68]" started-at="2022-11-23T00:01:52 ICT" name="Button_Detail" description="Clik the button more detail on slider" finished-at="2022-11-23T00:01:54 ICT" duration-ms="1364" status="PASS">
           <reporter-output>
           </reporter-output>
-        </test-method> <!-- DoLogin -->
-        <test-method is-config="true" signature="suiteTearDown()[pri:0, instance:test.Han.TestHan@1bd4fdd]" started-at="2022-11-22T11:47:34 ICT" name="suiteTearDown" finished-at="2022-11-22T11:47:35 ICT" duration-ms="308" status="PASS">
+        </test-method> <!-- Button_Detail -->
+        <test-method is-config="true" signature="suiteTearDown()[pri:0, instance:test.Khanh.TestSlider@4f83df68]" started-at="2022-11-23T00:01:54 ICT" name="suiteTearDown" finished-at="2022-11-23T00:01:55 ICT" duration-ms="1013" status="PASS">
           <reporter-output>
           </reporter-output>
         </test-method> <!-- suiteTearDown -->
-      </class> <!-- test.Han.TestHan -->
+      </class> <!-- test.Khanh.TestSlider -->
     </test> <!-- Default test -->
   </suite> <!-- Default suite -->
 </testng-results>
